third_party/WebKit/LayoutTests/fast/workers/storage/empty-statement-sync-expected.txt

CONSOLE WARNING: 'openDatabaseSync' is deprecated. Please switch to Indexed Database API.
This test tries to execute an empty statement.
PASS: result is null
PASS
